feat: SoundLoader の基底クラスを追加
